Move global ErrorBoundary to _app.tsx
